fixed the slowdown with the collision reporter, stopped the game from lagging so much when firing events for trigger regions (it still fires too many, but hey it's only one every few seconds now as opposed to every frame), and now we aren't looping through 1800+ manifolds every frame.
Also added a couple other speed fixes that the profiler was telling me about

git-svn-id: svn://svn.code.sf.net/p/ponykart/code/trunk@398 ba97e771-f6de-4074-9afc-763755853dd4

+++ Ponykart/Actors/Components/BillboardSetComponent.cs
@@ -8,7 +8,7 @@ namespace Ponykart.Actors {
 	/// Represents a billboard set, aka a collection of billboards that share the same material, direction, etc.
 	/// </summary>
 	public class BillboardSetComponent : LDisposable {
-		public long ID { get; protected set; }
+		public uint ID { get; protected set; }
 		public string Name { get; protected set; }
 		public BillboardSet BillboardSet { get; protected set; }
